fix: rename files for windows compatibility
Based on PR #196 by @Wonderand
